game-service: refactoring
